requirements.txt
Django==4.0.5
python-dotenv==0.20.0
Pillow==9.2.0
